% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{connectToAquarius}
\alias{connectToAquarius}
\title{Connect to Aquarius server}
\usage{
connectToAquarius(
  username,
  password = keyring::key_get("aquarius", username),
  hostname = "https://aquarius.nps.gov/aquarius",
  publish_api_url = "https://aquarius.nps.gov/aquarius/Publish/v2"
)
}
\arguments{
\item{username}{The AQTS credentials username}

\item{password}{The AQTS credentials password}

\item{hostname}{A server name or IP address}

\item{publish_api_url}{URL for the Aquarius Publish API.}
}
\value{
Invisibly returns server response
}
\description{
Before you use other functions in this package to get data from Aquarius,
you must use this function to log in. If accessing NPS data, you must be on
the DOI network and you only need to provide the read-only username and
password.
}
\examples{
\dontrun{
keyring::key_set("aquarius", "aqreadonly")  # Save credentials in keyring - only need to do this once per user & computer
connectToAquarius("aqreadonly")  # Connect to Aquarius using password saved in keyring
}
}
