% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{disconnectFromAquarius}
\alias{disconnectFromAquarius}
\title{Disconnect from Aquarius server}
\usage{
disconnectFromAquarius()
}
\description{
Use this function to log out when you are done retrieving data.
}
\examples{
\dontrun{
disconnectFromAquarius()
}
}
