% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{getFolders}
\alias{getFolders}
\title{Get names of folders in Aquarius}
\usage{
getFolders(search_string)
}
\arguments{
\item{search_string}{Optional. Text to search for in folder names. Accepts regular expressions. Case insensitive.}
}
\value{
Character vector of folder names
}
\description{
Use this to see the names of the folders in Aquarius. Use \code{search_string} to
see only folder names containing that text.
}
\examples{
\dontrun{
getFolders()
getFolders("OLYM")
}
}
