% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{getLocationInfo}
\alias{getLocationInfo}
\title{Get information about locations in Aquarius}
\usage{
getLocationInfo(search_string, folder, include_subfolders = TRUE)
}
\arguments{
\item{search_string}{Optional. Text to search for in location names and identifiers. Accepts regular expressions. Case insensitive.}

\item{folder}{Optional. Limit results to locations in this folder. You must use the full folder name (e.g. "National Park Service.Klamath Network", \emph{not} "Klamath Network").}

\item{include_subfolders}{Ignored if \code{folder} not specified. Indicates whether results should include locations in subfolders.}
}
\value{
Tibble containing location information
}
\description{
Use this to see locations in Aquarius.
}
\examples{
\dontrun{
getLocationInfo()
}
}
