% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{getTimeSeries}
\alias{getTimeSeries}
\title{Get corrected time series data}
\usage{
getTimeSeries(ts_id, start, end)
}
\arguments{
\item{ts_id}{Time series identifier}

\item{start}{Optional. Only retrieve data after and including this date/time. Can be an R POSIXct object, or a character string in ISO 8601 format (YYYY-MM-DDThh:mm:ss+/-hh:mm)}

\item{end}{Optional. Only retrieve data after and including this date/time. Can be an R POSIXct object, or a character string in ISO 8601 format (YYYY-MM-DDThh:mm:ss+/-hh:mm)}
}
\value{
A list that includes timeseries data and metadata
}
\description{
Get corrected time series data
}
\examples{
\dontrun{
# Get timeseries data from Jan 1 2012 at midnight PST to Jan 1 2013 at midnight PST.
time_series <- getTimeSeries("Discharge.Working@Location", start = "2012-01-01T00:00:00−08:00", end = "2013-01-01T00:00:00−08:00")
}
}
