% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{getTimeSeriesByFolder}
\alias{getTimeSeriesByFolder}
\title{Get all corrected time series data in a folder containing multiple locations}
\usage{
getTimeSeriesByFolder(
  folder,
  parameters,
  labels,
  start,
  end,
  include_subfolders = TRUE
)
}
\arguments{
\item{folder}{Limit results to locations in this folder. You must use the full folder name (e.g. "National Park Service.Klamath Network", \emph{not} "Klamath Network")}

\item{parameters}{A character vector of parameters to include}

\item{labels}{A character vector of labels to include}

\item{start}{Optional. Only retrieve data after and including this date/time. Can be an R POSIXct object, or a character string in ISO 8601 format (YYYY-MM-DDThh:mm:ss+/-hh:mm)}

\item{end}{Optional. Only retrieve data after and including this date/time. Can be an R POSIXct object, or a character string in ISO 8601 format (YYYY-MM-DDThh:mm:ss+/-hh:mm)}

\item{include_subfolders}{Ignored if \code{folder} not specified. Indicates whether results should include locations in subfolders.}
}
\value{
A list of time series by location
}
\description{
Get all corrected time series data in a folder containing multiple locations
}
\examples{
\dontrun{
olym_clim_ts <- getTimeSeriesByFolder("OLYM_Climate", c("Air Temp", "Snow Depth"), c("Avg", "Depth"))
}
}
