% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_client.R
\name{timeseriesClient_configureProxy}
\alias{timeseriesClient_configureProxy}
\title{Auto-configures the proxy to route all requests through Fiddler}
\value{
None
}
\description{
This method configures the R proxy to route everything through Fiddler if it is running
}
\examples{
\dontrun{
Sys.setenv(http_proxy="http://localhost:8888") # Enables Fiddler capturing of traffic
Sys.setenv(http_proxy="") # Disables Fiddler proxying
}
}
