% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_client.R
\name{timeseriesClient_connect}
\alias{timeseriesClient_connect}
\title{Connects to an AQTS server}
\arguments{
\item{hostname}{A server name or IP address}

\item{username}{The AQTS credentials username}

\item{password}{The AQTS credentials password}
}
\description{
Once authenticated, all subsequent requests to the AQTS server will reuse the authenticated session
}
\examples{
\dontrun{
connect("localhost", "admin", "admin") # When running R on your AQTS app server
connect("myserver", "me", "mypassword") # Connect over the network
connect("https://myserver", "user", "letmein") # Connect to an AQTS server with HTTPS enabled
}
}
