% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_client.R
\name{timeseriesClient_formatIso8601}
\alias{timeseriesClient_formatIso8601}
\title{Formats a datetime in ISO 8601 format}
\arguments{
\item{datetime}{A datetime object}
}
\value{
The time in YYYY-MM-DDTHH:mm:SS.fffffZ format
}
\description{
Formats a datetime in ISO 8601 format
}
