% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_client.R
\name{timeseriesClient_getFieldVisits}
\alias{timeseriesClient_getFieldVisits}
\title{Gets field visits}
\arguments{
\item{locationIdentifier}{Optional LocationIdentifier filter}

\item{queryFrom}{Optional QueryFrom filter}

\item{queryTo}{Optional QueryTo filter}

\item{activityType}{Optional DiscreteMeasurementActivity filter}
}
\value{
The activities performed at the locations during the requested time range
}
\description{
Gets field visits activities
}
