% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_client.R
\name{timeseriesClient_getLocationIdentifier}
\alias{timeseriesClient_getLocationIdentifier}
\title{Gets the location identifier from a time series identifier string}
\arguments{
\item{timeSeriesIdentifier}{A time-series identifier in \<Parameter\>.\<Label\>\@\<LocationIdentifier\> syntax}
}
\value{
The identifier of the location
}
\description{
Gets the location identifier from a time series identifier string
}
\examples{
\dontrun{
getLocationIdentifier("Stage.Working@MyLocation") # MyLocation
}
}
