% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_client.R
\name{timeseriesClient_getMetadataChangeTransactionList}
\alias{timeseriesClient_getMetadataChangeTransactionList}
\title{Gets Change list for a given time-series}
\arguments{
\item{timeSeriesIdentifier}{The time-series identifier or unique ID}

\item{queryFrom}{Optional QueryFrom filter}

\item{queryTo}{Optional QueryTo filter}
}
\value{
The list of change trasactions
}
\description{
Gets Change list for a given time-series
}
