% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_client.R
\name{timeseriesClient_getRatingModelOutputValues}
\alias{timeseriesClient_getRatingModelOutputValues}
\title{Gets output values from a rating model}
\arguments{
\item{ratingModelIdentifier}{The identifier of the rating model}

\item{inputValues}{The list of input values to run through the model}

\item{effectiveTime}{Optional time of applicability. Assumes current time if omitted}

\item{applyShifts}{Optional boolean, defaults to FALSE}
}
\value{
The output values from the applicable curve of the rating model. An output value of NA is returned if the input is outside the curve.
}
\description{
Gets output values from a rating model
}
