% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_client.R
\name{timeseriesClient_getRatings}
\alias{timeseriesClient_getRatings}
\title{Gets the rating models matching the optional filters}
\arguments{
\item{locationIdentifier}{Optional LocationIdentifier filter}

\item{queryFrom}{Optional QueryFrom filter}

\item{queryTo}{Optional QueryTo filter}

\item{inputParameter}{Optional InputParameter filter}

\item{outputParameter}{Optional OutputParameter filter}
}
\value{
A list of rating models and their applicable curves
}
\description{
Retrieves the rating models and their applicable curves
}
\examples{
\dontrun{
# Get all the ratings in effect during October 2016 at a location
ratings = timeseries$getRatings("A015001", "2016-10-01", "2016-10-31")
ratings$Identifier
ratings$Curves$Type
}
}
