% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_client.R
\name{timeseriesClient_getTimeSeriesCorrectedData}
\alias{timeseriesClient_getTimeSeriesCorrectedData}
\title{Get corrected data for a time-series}
\arguments{
\item{timeSeriesIdentifier}{The time-series identifier or unique ID}
}
\value{
The corrected data and metadata for the time-series
}
\description{
The getTimeSeriesData() method is usually a better choice, since it can pull corrected data from multiple time-series.
But when you need to look at the metadata of a time-series, this method is required.
}
