% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_client.R
\name{timeseriesClient_getTimeSeriesDescriptions}
\alias{timeseriesClient_getTimeSeriesDescriptions}
\title{Fetch all requested time-series descriptions matching the filter}
\arguments{
\item{locationIdentifier}{Optional location identifier filter}

\item{parameter}{Optional parameter filter}

\item{publish}{Optional publish filter}

\item{computationIdentifier}{Optional computation identifier filter}

\item{computationPeriodIdentifier}{Optional computation period identifier filter}

\item{extendedFilters}{Optional extended attribute filter}
}
\value{
All the time-series descriptions matching the filters
}
\description{
Fetch all requested time-series descriptions matching the filter
}
