% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_client.R
\name{timeseriesClient_getTimeSeriesUniqueId}
\alias{timeseriesClient_getTimeSeriesUniqueId}
\title{Gets the unique ID of a time-series from its identifier string}
\arguments{
\item{timeSeriesIdentifier}{A time-series identifier in \<Parameter\>.\<Label\>\@\<LocationIdentifier\> syntax}
}
\value{
The unique ID of the time-series
}
\description{
If the input string is already a unique ID, the input value is simply returned unmodified.
}
\examples{
\dontrun{
getTimeSeriesUniqueId("Stage.Working@MyLocation") # cdf184928c8249abb872f852f0fa7d01
}
}
