% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_client.R
\name{timeseriesClient_getUtcOffsetText}
\alias{timeseriesClient_getUtcOffsetText}
\title{Gets the UTC offset string from a numeric UTC offset hours value}
\arguments{
\item{utcOffset}{A UTC offset in hours}
}
\value{
An ISO8601 UTC offset string in +/-HH:MM format
}
\description{
Gets the UTC offset string from a numeric UTC offset hours value
}
\examples{
\dontrun{
getUtcOffsetText(-8) # "-08:00"
getUtcOffsetText(2.5) # "+02:30"
}
}
