% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_client.R
\name{timeseriesClient_parseIso8601}
\alias{timeseriesClient_parseIso8601}
\title{Parse an ISO 8601 timestamp into a POSIXct value}
\arguments{
\item{isoText}{An ISO 8601 timestamp string}
}
\value{
The equivalent POSIXct datetime
}
\description{
Parse an ISO 8601 timestamp into a POSIXct value
}
