% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_client.R
\name{timeseriesClient_toJSV}
\alias{timeseriesClient_toJSV}
\title{Converts an item to JSV format, for GET request query parameter values}
\description{
Converts vectors or named lists to JSV. Everything else is left unmodified.

Query parameters in a GET request need to be in JSV format.
JSON body parameters in POST/PUT/DELETE requests do not need JSV formatting (they are, JSON)

https://github.com/ServiceStack/ServiceStack.Text/wiki/JSV-Format
}
