% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_client.R
\name{timeseriesClient_uploadExternalReport}
\alias{timeseriesClient_uploadExternalReport}
\title{Uploads a file to a location as an external report}
\arguments{
\item{locationDataOrIdentifier}{Either a location identifier string, or a LocationData object from a previous getLocationData request}

\item{path}{The path to the file to be uploaded}

\item{title}{The title of the report to display in AQTS}

\item{deleteDuplicateReports}{If TRUE or missing, any existing reports with the same title will be deleted before the new report is uploaded}
}
\value{
The JSON response from a successful upload
}
\description{
Any existing report on AQTS with the same title will be
}
