% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{getTimeSeriesByLocation}
\alias{getTimeSeriesByLocation}
\title{Get all corrected time series data at a location}
\usage{
getTimeSeriesByLocation(loc_ids, parameters, labels, start, end)
}
\arguments{
\item{loc_ids}{A character vector of location identifiers}

\item{parameters}{A character vector of parameters to include}

\item{labels}{A character vector of labels to include}

\item{start}{Optional. Only retrieve data after and including this date/time. Can be an R POSIXct object, or a character string in ISO 8601 format (YYYY-MM-DDThh:mm:ss+/-hh:mm)}

\item{end}{Optional. Only retrieve data after and including this date/time. Can be an R POSIXct object, or a character string in ISO 8601 format (YYYY-MM-DDThh:mm:ss+/-hh:mm)}
}
\value{
A list of time series by location
}
\description{
Get all corrected time series data at a location
}
\examples{
\dontrun{
olym_clim_ts <- getTimeSeriesByLocation("OLYM_MET_DPC", c("Air Temp", "Snow Depth"), c("Avg", "Depth"))
}
}
