% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{getTimeSeriesInfo}
\alias{getTimeSeriesInfo}
\title{Get list of datasets at a location}
\usage{
getTimeSeriesInfo(loc_ids)
}
\arguments{
\item{loc_ids}{Character vector of location identifier(s)}
}
\value{
A tibble with one line per dataset
}
\description{
Get list of datasets at a location
}
\examples{
\dontrun{
getDatasetInfo("MORA_SUP")
}
}
